pages: splitted settings page
